requirements.txt
PySide6>=6.6.0
numpy>=1.24.0
Pillow>=10.0.0
pillow-heif>=0.16.0
pyinstaller>=6.5.0
requests>=2.31.0
pyproj>=3.6.0
matplotlib>=3.8.0
platformdirs>=4.0.0
